Apply hostbuild patches for on device builds as well
But do not run hostbuild step.

+++ scripts/build/termux_step_handle_hostbuild.sh
@@ -8,6 +8,13 @@ termux_step_handle_hostbuild() {
 		test -f "$patch" && sed "s%\@TERMUX_PREFIX\@%${TERMUX_PREFIX}%g" "$patch" | patch --silent -p1
 	done
 
+	if [ "$TERMUX_ON_DEVICE_BUILD" = "true" ]; then
+		# We do not need to hostbuild on device since we are
+		# not cross-compiling
+		return
+	fi
+
+	local TERMUX_HOSTBUILD_MARKER="$TERMUX_PKG_HOSTBUILD_DIR/TERMUX_BUILT_FOR_$TERMUX_PKG_VERSION"
 	if [ ! -f "$TERMUX_HOSTBUILD_MARKER" ]; then
 		rm -Rf "$TERMUX_PKG_HOSTBUILD_DIR"
 		mkdir -p "$TERMUX_PKG_HOSTBUILD_DIR"
